Add files via upload
In the world of data, I've harnessed Tableau's magic wand to conjure a mesmerizing array of visualizations. From intricate scatter plots that reveal hidden correlations to dynamic bar charts that dance with insights, each visualization is a symphony of colors and patterns that unveils the stories within the data. Whether it's mapping Starbucks locations, dissecting Walmart's sales, or exploring the vastness of data landscapes, Tableau has been my trusty brush, painting the canvas of insights. Dive into this kaleidoscope of visual wonders, and let the data unfold its secrets in a breathtaking display of analytical artistry.